Safari-609.1.11.1/LayoutTests/fast/dom/br-set-outerText-expected.txt

Tests that br.outerText does not throw

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS br.outerText = 'PASS'; did not throw exception.
PASS br.parentNode is null
PASS target.firstChild.data is "PASS"
PASS successfullyParsed is true

TEST COMPLETE
PASS
